fix(drivers): accuracy and hardware fixes across the emulation layer
Super Pac-Man / Pac & Pal sprite priority layer and flipscreen.
Centipede and Millipede trackball handling plus start-button lamps.
Pac-Man start lamps at 0x5004/0x5005; Make Trax protection patch moved
to an opcode-base image so the boot ROM checksum still passes.
Eyes and Ponpoko decrypt through the new rom_decrypt hook.
Mappy video orientation and scroll documented.
Assorted driver, sound and machine fixes, plus removal of the
now-redundant windows.h/Vulkan boundary guard blocks.

+++ aae/aae/vidhrdwr/vertigo_video.cpp
@@ -589,7 +589,7 @@ void vertigo_vproc(int cycles, int irq4)
              cut and run for just 100 cycles to make sure the
              loop is actually entered.
             */
-			if ((cmc->jpos != irq4) && cycles > 100)
+			if ((cmc->jpos != (uint32_t)irq4) && cycles > 100)
 			{
 				cycles=100;
 			}
